Task/odex 96 (#522)
* refactor: remove duplicate hook

* chore: remove component menu v2

* chore: remove redundant code v2
