--- src/components/Badge.module.scss
@@ -1,15 +1,0 @@
-@import 'src/styles/mixins';
-@import 'src/styles/themes';
-
-.badge {
-  $height: 21px;
-  @include flex;
-  display: inline-flex;
-  height: $height;
-  border-radius: calc($height / 2);
-  background: fade-out(white, 0.9);
-  color: $white-color;
-  font-size: 10px;
-  font-weight: 500;
-  min-width: 90px;
-}
