--- src/components/CenterEllipsis.tsx
@@ -1,27 +1,0 @@
-import React, { useState } from 'react';
-import Tooltip from './Tooltip';
-
-const CenterEllipsis: React.FC<{
-  text: string | String;
-  size: number;
-  className?: string;
-}> = ({ text, size, className }) => {
-  const [content, setContent] = useState(text);
-
-  return (
-    <Tooltip
-      content={content}
-      onUntrigger={() => {
-        setContent(text);
-      }}
-      maxWidth={'none'}
-    >
-      <span className={className} style={{ overflow: 'hidden', marginBottom: 0 }}>{`${text.slice(
-        0,
-        size
-      )}...${text.slice(-size)}`}</span>
-    </Tooltip>
-  );
-};
-
-export default CenterEllipsis;
