--- src/components/connect-wallet/ConnectWallet.module.scss
@@ -1,8 +1,0 @@
-.options {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;  
-}
-
-
-
