--- src/components/connect-wallet/ConnectWalletModal.module.scss
@@ -1,94 +1,0 @@
-.select {
-  background: #141416;
-  border-radius: 24px;
-  padding: 27px 24px;
-
-  .title {
-      display: flex;
-      align-items: flex-end;
-
-      gap: 5px;
-      margin-bottom: 25px;
-      font-family: IBM Plex Sans;
-      font-style: normal;
-      font-weight: 600;
-      font-size: 18px;
-      line-height: 100%;
-      /* identical to box height, or 18px */
-
-      text-align: center;
-
-      /* aiRight/White */
-
-      color: #ffffff;
-  }
-
-  .options {
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-
-      .item {
-          background: #1e1e21;
-          border-radius: 8px;
-          display: flex;
-          gap: 10px;
-          align-items: center;
-          padding: 16px;
-
-          &:hover {
-              cursor: pointer;
-              background: #333642 !important;
-          }
-
-          .logo {
-              width: 54px;
-              height: 54px;
-          }
-
-          .grow {
-              flex-grow: 1;
-
-              display: flex;
-              flex-direction: column;
-              gap: 10px;
-              .network-title {
-                  font-family: IBM Plex Sans;
-                  font-style: normal;
-                  font-weight: 600;
-                  font-size: 18px;
-                  line-height: 100%;
-                  /* identical to box height, or 18px */
-
-                  /* aiRight/White */
-
-                  color: #ffffff;
-              }
-
-              .des {
-                  font-family: IBM Plex Sans;
-                  font-style: normal;
-                  font-weight: 500;
-                  font-size: 14px;
-                  line-height: 150%;
-                  /* identical to box height, or 21px */
-
-                  font-feature-settings: "ss03" on;
-
-                  /* aiRight/Gray 3 */
-
-                  color: #777e90;
-              }
-          }
-
-          .arrow-right {
-              width: 0;
-              height: 0;
-              border-top: 8px solid transparent;
-              border-bottom: 8px solid transparent;
-
-              border-left: 8px solid #ffffff;
-          }
-      }
-  }
-}
