--- src/components/connect-wallet/LoginWidget.module.scss
@@ -1,100 +1,0 @@
-@import 'src/styles/variables';
-@import 'src/styles/mixins';
-
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 60px;
-
-  iframe {
-    border: none;
-    width: 200px;
-    height: 100%;
-  }
-}
-
-.address {
-  color: $white;
-  margin: 0 10px 0 8px;
-}
-
-button.connected {
-  border: 1px solid $white;
-  text-transform: none;
-}
-
-button.connect {
-  text-transform: none;
-
-  img {
-    margin-right: 10px;
-  }
-}
-
-.item {
-  background: #2d2938;
-  border-radius: 8px;
-  display: flex;
-  gap: 12px;
-  align-items: center;
-  padding: 8px;
-
-  //light mode
-  &.light {
-    background-color: #efefef;
-  }
-
-  &:last-child {
-    margin-bottom: 28px;
-  }
-
-  &:hover {
-    @include theme() {
-      background-color: theme-get('bg-btn-primary');
-    }
-  }
-
-  &.light:hover .des {
-    color: #efefef !important;
-  }
-
-  .logo {
-    box-sizing: border-box;
-    width: 36px;
-    height: 36px;
-    border-radius: 8px;
-  }
-
-  .grow {
-    flex-grow: 1;
-
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    
-    .des {
-      font-style: normal;
-      font-weight: 500;
-      font-size: 14px;
-      line-height: 150%;
-      /* identical to box height, or 21px */
-      font-feature-settings: 'ss03' on;
-      user-select: none;
-      color: #bfa7f2;
-
-      //light mode
-      &.light {
-        color: #6f767e;
-      }
-    }
-  }
-
-  .arrow-right {
-    width: 0;
-    height: 0;
-    border-top: 8px solid transparent;
-    border-bottom: 8px solid transparent;
-    border-left: 8px solid #ffffff;
-  }
-}
