--- src/hooks/useClickOutSide.ts
@@ -1,24 +1,0 @@
-import { useRef, useEffect } from 'react';
-
-const useClickOutside = (callback) => {
-  const ref = useRef();
-
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      //@ts-ignore
-      if (ref.current && !ref.current.contains(event.target)) {
-        callback();
-      }
-    };
-
-    document.addEventListener('click', handleClickOutside);
-
-    return () => {
-      document.removeEventListener('click', handleClickOutside);
-    };
-  }, [callback]);
-
-  return ref;
-};
-
-export default useClickOutside;
