--- src/layouts/MenuV3.module.scss
@@ -1,314 +1,0 @@
-@import 'src/styles/mixins';
-@import 'src/styles/themes';
-
-.menu {
-  display: flex;
-  justify-content: space-between;
-  padding: 20px 120px;
-  position: fixed;
-  width: 100%;
-  top: 0;
-  z-index: 999;
-
-  .logo {
-    display: flex;
-  }
-
-  @include theme() {
-    background-color: theme-get('neutral-surface-bg');
-    border-bottom: 1px solid theme-get('neutral-border-border-default');
-  }
-}
-
-.menuMobile {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  padding: 16px;
-  z-index: 999;
-  position: fixed;
-  top: 0;
-  left: 0;
-
-  &_text {
-    font-weight: 600;
-    font-size: 24px;
-    line-height: 100%;
-    margin-left: 10px;
-    color: #fff;
-  }
-
-  .logo {
-    display: flex;
-    align-items: center;
-
-    svg {
-      margin-right: 10px;
-    }
-  }
-
-  @include theme() {
-    background-color: theme-get('neutral-surface-bg');
-    border-bottom: 1px solid theme-get('neutral-border-border-default');
-  }
-}
-
-.sideMenu {
-  margin-left: -70%;
-  width: 70%;
-  height: 100vh;
-  padding: 40px 24px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  z-index: $zindex-sticky;
-  overflow: auto;
-  transition: margin-left 0.2s ease-in-out;
-
-  &.open {
-    margin-left: 0;
-  }
-
-  @include theme() {
-    background-color: theme-get('bg-color');
-    border-bottom: 1px solid theme-get('neutron-5');
-  }
-
-  .menu_list {
-    flex-direction: column;
-  }
-}
-
-.menu_others_list {
-  min-height: 100px;
-  width: 200px;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  border-radius: 6px;
-
-  padding: 16px 13px 16px 16px;
-  color: $white-color;
-  background-color: $dropdown-bg-color-dark;
-  border: 1px solid $dropdown-border-color-dark;
-
-  .menu_item {
-    display: flex;
-    height: 48px;
-    padding: 0px 6px;
-    align-items: center;
-    gap: 8px;
-    align-self: stretch;
-    border-radius: 8px;
-
-    &:hover {
-      background: $dropdown-bg-color-hover-dark;
-    }
-  }
-
-  &.light {
-    color: $dropdown-color-light;
-    background-color: $white-color;
-    border: 1px solid $dropdown-border-color-light;
-
-    .menu_item:hover {
-      background: $dropdown-bg-color-hover-light;
-    }
-  }
-}
-
-.menuLeft {
-  display: flex;
-  align-items: center;
-
-  .divider {
-    @include theme() {
-      border: 1px solid theme-get('neutral-border-border-bold');
-    }
-    margin: 0 7px 0 24px;
-    height: 40px;
-  }
-}
-
-.menu_list {
-  display: flex;
-
-  .menu_item {
-    white-space: nowrap;
-    box-sizing: border-box;
-    position: relative;
-    vertical-align: middle;
-    display: flex;
-    align-items: center;
-    border-radius: 8px;
-    padding: 6px 10px;
-    margin-left: 17px;
-    margin-right: 17px;
-    cursor: pointer;
-
-    &::before,
-    &::after {
-      box-sizing: inherit;
-      content: '';
-      position: absolute;
-      top: 8px;
-      width: 105%;
-      height: 72%;
-    }
-
-    .menu_item_text {
-      text-transform: uppercase;
-      font-weight: 500;
-    }
-
-    &:hover {
-      .menu_item_text {
-        @include theme() {
-          color: theme-get('neutral-1');
-        }
-      }
-    }
-
-    &.active {
-      @include theme() {
-        color: theme-get('neutral-1');
-      }
-
-      span {
-        @include theme() {
-          color: theme-get('neutral-1');
-        }
-      }
-
-      &::before {
-        @include theme() {
-          border-left-color: theme-get('neutral-1');
-          border-right-color: theme-get('neutral-1');
-          border-bottom-color: theme-get('neutral-1');
-        }
-      }
-
-      &::after {
-        border: 0 solid transparent;
-      }
-
-      &::after {
-        transform: rotate(240deg);
-      }
-    }
-
-    &_text {
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 100%;
-      justify-items: center;
-      display: contents;
-
-      @include theme() {
-        color: theme-get('neutral-3');
-      }
-
-      &.active {
-        @include theme() {
-          color: theme-get('neutral-1');
-        }
-      }
-    }
-
-    &_others {
-      display: flex;
-    }
-
-    @include mobile() {
-      margin-bottom: 24px;
-      margin-left: 0;
-      margin-right: 0;
-      width: fit-content;
-    }
-  }
-}
-
-.connect_wallet_wrapper {
-  display: flex;
-  align-items: center;
-
-  .menu_theme {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 28px;
-    padding: 9px;
-    width: 45px;
-    height: 45px;
-    margin-right: 20px;
-    box-shadow: 0px 3.33333px 6.66667px -3.33333px rgba(0, 0, 0, 0.25),
-      inset 0px -0.833333px 0.833333px rgba(0, 0, 0, 0.49), inset 0px 1.66667px 0.833333px rgba(255, 255, 255, 0.06);
-
-    &:hover {
-      background: #2d2938;
-      cursor: pointer;
-    }
-
-    &.active {
-      background: #2d2938;
-
-      &.light {
-        background-color: #fcfcfc;
-        box-shadow: 0px 1px 6px -2px rgba(0, 0, 0, 0.25), inset 0px -0.833333px 0.833333px rgba(0, 0, 0, 0.04),
-          inset 0px 1.66667px 0px rgba(255, 255, 255, 0.25);
-      }
-    }
-
-    &_text {
-      font-style: normal;
-      font-weight: 500;
-      font-size: 14px;
-      line-height: 100%;
-      margin-left: 10px;
-    }
-  }
-}
-
-.spin {
-  border-radius: 86%;
-  box-shadow: none;
-
-  &::before,
-  &::after {
-    top: 0;
-    left: -2px;
-    border-radius: 50%;
-    transform: rotate(-8.549deg);
-  }
-
-  &::before {
-    border: 1px solid transparent;
-  }
-
-  &:active::before {
-    @include theme() {
-      border-left-color: theme-get('neutral-1');
-      border-right-color: theme-get('neutral-1');
-      border-bottom-color: theme-get('neutral-1');
-    }
-
-    transition: border-left-color 0.05s linear, border-right-color 0.05s linear 0.05s,
-      border-bottom-color 0.05s linear 0.05s;
-  }
-
-  &::after {
-    border: 0 solid transparent;
-  }
-
-  &:active::after {
-    border-bottom-width: 1px;
-    border-left-width: 1px;
-    border-right-width: 1px;
-    transform: rotate(240deg);
-    transition: transform 0.1s linear 0s, border-left-width 0s linear 0.25s;
-  }
-}
